fix(reliability): stop /api/health-check-cron from HTTP-fetching itself
health-check-cron made two outbound fetch() calls to this same server's
own /api/health and /api/readiness routes to build its report. That
self-referential network round-trip added avoidable latency and, under
concurrent load, contended with the server's own limited request
capacity for no benefit over calling the logic directly.

Extracted /api/health's report-building logic into
lib/health/report.ts (buildHealthReport()), unchanged in behavior — the
route handler now just calls it and returns the same JSON shape.
health-check-cron now calls buildHealthReport() and
getDeploymentReadiness() in-process instead of fetching its own URLs.

Verified: existing tests/integration/api/health-route.test.ts passes
unmodified (mocks target the same underlying lib modules regardless of
which file imports them), full suite (2895 tests) unaffected, and
manually confirmed live that both routes return the same
ok/degraded/503 shape as before.
